+++ core/common/src/main/java/alluxio/Constants.java
@@ -182,11 +182,11 @@ public final class Constants {
   public static final String KERBEROS_DEFAULT_AUTH_TO_LOCAL = "DEFAULT";
 
   // Encryption
-  public static final int DEFAULT_BLOCK_HEADER_SIZE = 0;
-  public static final int DEFAULT_BLOCK_FOOTER_SIZE = 0;
-  public static final int DEFAULT_CHUNK_HEADER_SIZE = 0;
-  public static final int DEFAULT_CHUNK_SIZE = 64 * KB;
-  public static final int DEFAULT_CHUNK_FOOTER_SIZE = 16;
+  public static final long DEFAULT_BLOCK_HEADER_SIZE = 0L;
+  public static final long DEFAULT_BLOCK_FOOTER_SIZE = 0L;
+  public static final long DEFAULT_CHUNK_HEADER_SIZE = 0L;
+  public static final long DEFAULT_CHUNK_SIZE = 64 * KB;
+  public static final long DEFAULT_CHUNK_FOOTER_SIZE = 16L;
   public static final long INVALID_ENCRYPTION_ID = -1L;
 
   // ALLUXIO CS END
